refactor(cli): rename the misnamed `dev_vz` module to `builder_vm`
`commands/env/dev_vz` has nothing to do with the (removed) Vz backend — the
name is a leftover from when the dev VM was Vz-based. As its own module doc
says, it is the builder-VM image bootstrap module (Stage 0, workload-kernel
build, and bundled-image fetching). Rename the module file + submodule
directory to `builder_vm` and update all references across the workspace.

Pure rename; no behavior change. `cargo check --all-features`, clippy, fmt,
and the workspace tests are clean.
