+++ crates/mvm-cli/src/commands/image/mod.rs
@@ -476,7 +476,7 @@ fn resolve_or_pull_run_image_with(
     // ext4 writer cannot faithfully emit a tree. `mvmctl image pull` reaps
     // previous helper processes at startup; the run-image path is the other
     // place a builder VM can spawn, so give it the same sweep before we might add one.
-    crate::commands::env::dev_vz::sweep_orphaned_vm_helpers_on_startup();
+    crate::commands::env::builder_vm::sweep_orphaned_vm_helpers_on_startup();
 
     // Local sources route to their own ingest; a registry reference falls
     // through to the cache-or-pull path below.
