+++ crates/mvm/src/vm/reconcile.rs
@@ -19,7 +19,7 @@
 //! (which spawn subprocesses), never a VM boot, never Nix. The heavier
 //! helper-PID argv sweep stays in `mvmctl cache prune` (on by default). This
 //! reuses the live-vs-orphan discrimination from `mvm-cli`'s
-//! `env::dev_vz` reaper rather than reinventing the policy; the bare
+//! `env::builder_vm` reaper rather than reinventing the policy; the bare
 //! syscall is restated here only because the lower `mvm` crate can't depend on
 //! `mvm-cli`.
